requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
supabase
pydantic
pydantic-settings
python-multipart
